+++ src/app/public-portal/public-user-registration/add-public-user/add-public-user.component.ts
@@ -583,9 +583,13 @@ export class AddPublicUserComponent implements OnInit {
       this.authorizeRequestService.saveCitizenAndFormData(this.documentList, this.citizenDTO)
         .subscribe((result) => {
           if (result && this.paymentMethod !== PaymentMethod.ONLINE) {
+            let tempDataId = this.tokenStorageService.getFormData(this.tokenStorageService.CITIZEN_REGISTRATION_KEY);
+            if (tempDataId) {
+              this.authorizeRequestService.deleteTempData(tempDataId).subscribe();
+              this.tokenStorageService.removeFormData(this.tokenStorageService.CITIZEN_REGISTRATION_KEY);
+            }
             this.snackBar.success(this.systemService.getTranslation('ALERT.MESSAGE.REGISTRATION_SUCCESS'));
             this.router.navigate(['/login']);
-            this.tokenStorageService.removeFormData(this.tokenStorageService.CITIZEN_REGISTRATION_KEY);
           } else {
             this.snackBar.error('Operation failed');
           }
